Segreagate the Login and Register file Components and move both files to Screen folder for better code readibilty and reuse
