+++ Client/teams/Navigation/StackNavigator.tsx
@@ -1,14 +1,16 @@
 import React from 'react';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import {NavigationContainer} from '@react-navigation/native';
-import Testing from '../Screens/testing';
 import Home from '../Components/Home';
 import TeamsSplashscreen from '../Screens/TeamsSplashscreen';
+import Register from '../Screens/Register';
+import Login from '../Screens/Login';
 
 export type RootStackParams = {
   Home: undefined;
-  testing: undefined;
   Splash: undefined;
+  Register: undefined;
+  Login: undefined;
 };
 
 const stack = createNativeStackNavigator<RootStackParams>();
@@ -22,12 +24,18 @@ const StackNavigator = () => {
           name="Splash"
           component={TeamsSplashscreen}
         />
-        <stack.Screen name="Home" component={Home} />
         <stack.Screen
           options={{headerShown: false}}
-          name="testing"
-          component={Testing}
+          name="Login"
+          component={Login}
         />
+        <stack.Screen
+          options={{headerShown: false}}
+          name="Register"
+          component={Register}
+        />
+
+        <stack.Screen name="Home" component={Home} />
       </stack.Navigator>
     </NavigationContainer>
   );
