--- Client/teams/Screens/testing.tsx
@@ -1,13 +1,0 @@
-import {View} from 'react-native';
-import React from 'react';
-import Login from './Login';
-
-const Testing = () => {
-  return (
-    <View style={{flex: 1}}>
-      <Login />
-    </View>
-  );
-};
-
-export default Testing;
